Game File Loading/Saving System - Refractor of the GFXLibrary and File
WARNING WIP - THERE WILL BE BUGS
MAJOR UPDATE
TODO: Try to reduce the file size - more the 7 times the game size is a bit much....

Added a saving and loading system for the ATD images. Now loads all files from the game and saves it to our project when used in the editor. It will save all ATD files to a .pck when it's a standalone build. That .pck will be loaded into the game for later use.

Added a loading scene, for picking the path to your ATD install.

Changed the main scene, from the base scene to the loading scene.
Added textures for testing purposes to the base scene.

Changed the system behind getting the textures from the GFXLibrary - now you have to use GFXFile.GetTexture() to retrieve a chacheble texture

Updated the README

Updated the background color to black, the window size to Full HD and the resizing method to stretch.