New builder interface and loader refractor -BROKEN-
You now import the image files from a new menu inside godot. Just enter the path to ATD (It tries to auto fill it from your settings, eg. prev runs) and hit the button. Loading takes some time. The images will not appear instantly in your project view, as godot seldom refreshes these.

This is BROKEN though. Godot will crash when you try to rebuild the mono assemblies. Godot struggles to handle the threads that are being used to build the images at a reasonable speed. This issues is fixed in the 3.2 update of godot though and the error will only occur, when you try to import the files.